#12844 [통합테스트][게시판관리] 게시글이 존재하는 게시판을 삭제할 경우 알림창 생성

+++ src/main/kotlin/co/brainz/itsm/board/service/BoardService.kt
@@ -169,7 +169,9 @@ class BoardService(
      */
     @Transactional
     fun deleteBoard(boardAdminId: String): Boolean {
-        boardAdminRepository.deleteById(boardAdminId)
+        if (!boardAdminRepository.existsById(boardAdminId)) {
+            boardAdminRepository.deleteById(boardAdminId)
+        }
         return !boardAdminRepository.existsById(boardAdminId)
     }
